Particle Effects attached to Hat in order to keep them from rotating.
The HatFollowBody script keeps the hat oriented in the direction the player is moving and keeps it from rotating with it's parent object. I'm using this functionality for certain Particle Effects' benefit, namely PE_CalmDown and PE_CantTouchThisPurple. The other particle effects simply originate from a position but those circle it. Using it this way allows the code to remain virtually the same, as well as providing the non-rotating benefits.