fixing bug where main menu sound wouldn't work until slider was pressed
Problem was caused when sound was supposed to be played before any PlayerPrefs settings were set, the issue was solved when pressing the slider because after changing it the SaveData function on the VolumeSlider would be called, effectively saving the setting.

Problem was solved by having a default value to fall on, if there is no PlayerPrefs for volume yet, the default value will be max volume.